+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Update styles for work and edition pages
 - Add login buttons and info blurb UP items
 - SFR-1869: Add assertion step to minimize timeout errors
+- Fix: Improve accessibility of CTAs and search bar
 
 ## [0.17.5]